Refactor.  Move types into different module.
